Moving logging to `mavis.test`
This reduces the amount of code we have in `conftest.py` and keeps the
code related to hooks in a single place.
